+++ common/src/main/java/org/infernalstudios/questlog/core/quests/display/ObjectiveDisplayData.java
@@ -16,6 +16,15 @@ public class ObjectiveDisplayData {
     private final Component name;
     @Nullable
     private Objective objective;
+    private int indentLevel = 0;
+
+    public int getIndentLevel() {
+        return this.indentLevel;
+    }
+
+    public void setIndentLevel(int indentLevel) {
+        this.indentLevel = indentLevel;
+    }
 
     public ObjectiveDisplayData(JsonObject data) {
         String name = JsonUtils.getOrDefault(data, "name", (String) null);