+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - UNRELEASED
+
+### Added
+
+- Added an in-game config screen.
+
+### Fixed
+
+- Fixed issues with quest progress not properly resetting.
+- Fixed nested objectives not displaying child objectives.
+
 ## [2.3.1] - 2026-06-02
 
 ### Fixed